+++ seer-core/src/main/scala/graphics/Viewport.scala
@@ -6,7 +6,7 @@ object Viewport {
 	def apply( l:Int, b:Int, w:Int, h:Int) = new Viewport(l,b,w,h)
 	def apply( v:Viewport ) = new Viewport(v.l,v.b,v.w,v.h)
 }
-class Viewport(var l:Int, var b:Int, var w:Int, var h:Int) {
+class Viewport(var l:Float, var b:Float, var w:Float, var h:Float) {
 	def aspect() = { if (h!=0 && w!=0) w*1.f/h else 1.f }
-	def set(ll:Int, bb:Int, ww:Int, hh:Int){ l=ll; b=bb; w=ww; h=hh; }
+	def set(ll:Float, bb:Float, ww:Float, hh:Float){ l=ll; b=bb; w=ww; h=hh; }
 }
